feat: clean fav

+++ lib/pages/fav_detail/view.dart
@@ -1,3 +1,4 @@
+import 'package:PiliPlus/common/widgets/dialog.dart';
 import 'package:PiliPlus/http/loading_state.dart';
 import 'package:PiliPlus/http/user.dart';
 import 'package:PiliPlus/models/user/fav_folder.dart';
@@ -159,23 +160,51 @@ class _FavDetailPageState extends State<FavDetailPage> {
                                       },
                                       child: Text('编辑信息'),
                                     ),
+                                    PopupMenuItem(
+                                      onTap: () {
+                                        UserHttp.cleanFav(mediaId: mediaId)
+                                            .then((data) {
+                                          if (data['status']) {
+                                            SmartDialog.showToast('清除成功');
+                                            _favDetailController.onReload();
+                                          } else {
+                                            SmartDialog.showToast(data['msg']);
+                                          }
+                                        });
+                                      },
+                                      child: Text('清除失效内容'),
+                                    ),
                                     if (!Utils.isDefault(
                                         _favDetailController.item.value.attr ??
                                             0))
                                       PopupMenuItem(
                                         onTap: () {
-                                          UserHttp.deleteFolder(
-                                              mediaIds: [mediaId]).then((data) {
-                                            if (data['status']) {
-                                              SmartDialog.showToast('删除成功');
-                                              Get.back(result: true);
-                                            } else {
-                                              SmartDialog.showToast(
-                                                  data['msg']);
-                                            }
-                                          });
+                                          showConfirmDialog(
+                                            context: context,
+                                            title: '确定删除该收藏夹?',
+                                            onConfirm: () {
+                                              UserHttp.deleteFolder(
+                                                      mediaIds: [mediaId])
+                                                  .then((data) {
+                                                if (data['status']) {
+                                                  SmartDialog.showToast('删除成功');
+                                                  Get.back(result: true);
+                                                } else {
+                                                  SmartDialog.showToast(
+                                                      data['msg']);
+                                                }
+                                              });
+                                            },
+                                          );
                                         },
-                                        child: Text('删除'),
+                                        child: Text(
+                                          '删除',
+                                          style: TextStyle(
+                                            color: Theme.of(context)
+                                                .colorScheme
+                                                .error,
+                                          ),
+                                        ),
                                       ),
                                   ],
                                 )
